btc-strategy/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
requests>=2.28.0
websockets>=12.0
schedule>=1.2.0
python-dotenv>=1.0.0
